feat: move ids from int to uint64

+++ internal/service/blueprints.go
@@ -10,7 +10,7 @@ import (
 	"card-watcher/internal/models"
 )
 
-func (s *service) ListBlueprints(ctx context.Context, expansionID int, name string) (*models.ListBlueprintsResponse, error) {
+func (s *service) ListBlueprints(ctx context.Context, expansionID uint64, name string) (*models.ListBlueprintsResponse, error) {
 	blueprints, err := s.cardtraderAdapter.GetBlueprints(ctx, expansionID)
 	if err != nil {
 		return nil, fmt.Errorf("getting blueprints from cardtrader adapter: %w", err)
@@ -23,17 +23,17 @@ func (s *service) ListBlueprints(ctx context.Context, expansionID int, name stri
 			// filter via card name
 			if strings.Contains(strings.ToLower(blueprint.Name), normalizedName) {
 				resultingBlueprints = append(resultingBlueprints, &models.Blueprint{
-					Id:          int32(blueprint.ID),
+					Id:          blueprint.ID,
 					Name:        blueprint.Name,
-					ExpansionId: int32(blueprint.ExpansionID),
+					ExpansionId: blueprint.ExpansionID,
 				})
 			}
 		} else {
 			// no filter provided, return all cards from given expansion
 			resultingBlueprints = append(resultingBlueprints, &models.Blueprint{
-				Id:          int32(blueprint.ID),
+				Id:          blueprint.ID,
 				Name:        blueprint.Name,
-				ExpansionId: int32(blueprint.ExpansionID),
+				ExpansionId: blueprint.ExpansionID,
 			})
 		}
 	}
